🗺️ Atlas: Splitting storage/redb_cold_storage.rs
